Update winzone.lua
Updates Win Zones to allow for changing any Win Zone into an inactive state, that then requires activation from "ifused" fields in Trigger Zones or collectable entities.

This is a response to the feature request for #3618

Usage:
1. Place a winzone on the map
2. click on the win zone and change its name from "Win Zone" to a custom name
3. Place a trigger zone on the map
4. click on the trigger zone and put the custom name of the win zone in the ifused field of the trigger zone

This will make it so the player is required to reach the trigger zone before he can exit the level at the win zone.

This script update does not affect how Win Zones previously work. If you keep them named "Win Zone" in the properties panel they will operate as normal.