Fix for Voiceless Male CC Characters
Defaults "male" as the voice/soundset for Character Creator (CC) made characters. Originally these characters would either be voiceless or female. 

Selecting a female attachment in the CC should still set the voice type to female. Which can also now be optionally changed within the entity properties in the editor.