Fixed Rotation Issue with Lightray Decal
This fixes the issue with the lightray decal which is intended to be a light ray coming out from a window with no rotation. 

Unlike other decals, this decal when it rotates to always face the player, breaks the illusion that its emitting from a fixed position. By removing the decal facing lua in the FPE, this fixes the light-ray's casting direction.

Essentially it now operates as it did in FPSC, which is to add atmosphere to areas with windows.