Create healthboost.lua
This script will boost the health by 50 points, even if the player's health is full. This is similar to the earliest GGC versions where collecting health items would exceed the health beyond the maximum. It appears a few users want this old system still around.

I've made a change to the Syringe item so that it references this script and acts as a booster rather than health recovery.