Update proximine.lua
This update will make it so the countdown timer sound effect doesn't ever play twice over. Which fixes a bug whereas the count down timer can sometimes be heard even after having exploded.

Additionally this update allows for the option of "IfUsed" on switches to detonate specified proximity mines on the map.