Create windoor.lua
New script that will allow for making an exit door to your level to advance to a new level once successfully opened.

Its possible to also make it so this door is locked, requiring a key, in order to open it and exit the level. By using the "Use Key" property and putting the name of the key that is required to be collected.